# server/src/api/models/ApiResultResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * intrawiki
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { ErrorCode } from './ErrorCode';
import {
    ErrorCodeFromJSON,
    ErrorCodeFromJSONTyped,
    ErrorCodeToJSON,
} from './ErrorCode';

/**
 * 
 * @export
 * @interface ApiResultResponse
 */
export interface ApiResultResponse {
    /**
     * 
     * @type {boolean}
     * @memberof ApiResultResponse
     */
    success: boolean;
    /**
     * 
     * @type {number}
     * @memberof ApiResultResponse
     */
    status: number;
    /**
     * 
     * @type {ErrorCode}
     * @memberof ApiResultResponse
     */
    code: ErrorCode;
    /**
     * 
     * @type {string}
     * @memberof ApiResultResponse
     */
    message?: string;
}

/**
 * Check if a given object implements the ApiResultResponse interface.
 */
export function instanceOfApiResultResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "success" in value;
    isInstance = isInstance && "status" in value;
    isInstance = isInstance && "code" in value;

    return isInstance;
}

export function ApiResultResponseFromJSON(json: any): ApiResultResponse {
    return ApiResultResponseFromJSONTyped(json, false);
}

export function ApiResultResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiResultResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'success': json['success'],
        'status': json['status'],
        'code': ErrorCodeFromJSON(json['code']),
        'message': !exists(json, 'message') ? undefined : json['message'],
    };
}

export function ApiResultResponseToJSON(value?: ApiResultResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'success': value.success,
        'status': value.status,
        'code': ErrorCodeToJSON(value.code),
        'message': value.message,
    };
}