# server/src/api/models/AuthInfoResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * intrawiki
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { UserType } from './UserType';
import {
    UserTypeFromJSON,
    UserTypeFromJSONTyped,
    UserTypeToJSON,
} from './UserType';

/**
 * 
 * @export
 * @interface AuthInfoResponse
 */
export interface AuthInfoResponse {
    /**
     * 
     * @type {boolean}
     * @memberof AuthInfoResponse
     */
    loggedIn: boolean;
    /**
     * 
     * @type {number}
     * @memberof AuthInfoResponse
     */
    userId?: number;
    /**
     * 
     * @type {string}
     * @memberof AuthInfoResponse
     */
    account?: string;
    /**
     * 
     * @type {UserType}
     * @memberof AuthInfoResponse
     */
    userType?: UserType;
}

/**
 * Check if a given object implements the AuthInfoResponse interface.
 */
export function instanceOfAuthInfoResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "loggedIn" in value;

    return isInstance;
}

export function AuthInfoResponseFromJSON(json: any): AuthInfoResponse {
    return AuthInfoResponseFromJSONTyped(json, false);
}

export function AuthInfoResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthInfoResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'loggedIn': json['loggedIn'],
        'userId': !exists(json, 'userId') ? undefined : json['userId'],
        'account': !exists(json, 'account') ? undefined : json['account'],
        'userType': !exists(json, 'userType') ? undefined : UserTypeFromJSON(json['userType']),
    };
}

export function AuthInfoResponseToJSON(value?: AuthInfoResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'loggedIn': value.loggedIn,
        'userId': value.userId,
        'account': value.account,
        'userType': UserTypeToJSON(value.userType),
    };
}